requirements.txt
torch>=1.9.0
torchvision>=0.10.0
datasets>=2.0.0
transformers>=4.20.0
pillow>=8.3.0
numpy>=1.21.0
matplotlib>=3.4.0
tqdm>=4.62.0
opencv-python>=4.5.0
albumentations>=1.3.0
wandb>=0.12.0
accelerate>=0.20.0
